Remove pipeline artifacts.
These should never have been checked in, but in doing so they cause
subtle errors in the pipeline processing when used in conjunction with
the repository-agent app. These files are created as part of running the
pipeline, and used as the basis for subsequent steps. However, if the
repository agent is regularly updating them, then the repo versions are
used, instead of the pipeline-produced-versions, so out of date schemas
and data are used for some steps.

Added .gitignore to ignore these artifacts in the future.